refactor(staking): use DriftBanner in NotificationBanners component
